dedupe some check class names
