+++ MapsetVerifier.Checks.Tests/Mania/Skinning/CheckSkinningManiaTests.cs
@@ -1,10 +1,10 @@
-using MapsetVerifier.Checks.Taiko.Skinning;
+using MapsetVerifier.Checks.Mania.Skinning;
 using MapsetVerifier.Framework.Objects;
 using Xunit;
 
-namespace MapsetVerifier.Checks.Tests.Taiko.Skinning;
+namespace MapsetVerifier.Checks.Tests.Mania.Skinning;
 
-public class CheckSkinningTests
+public class CheckSkinningManiaTests
 {
     private static string BuildMinimalOsu(string hitObject = "256,192,1000,1,0,0:0:0:0:") =>
         string.Join(
@@ -12,7 +12,7 @@ public class CheckSkinningTests
             "osu file format v14",
             "[General]",
             "AudioFilename: audio.mp3",
-            "Mode: 1",
+            "Mode: 3",
             "[Metadata]",
             "Title:Title",
             "Artist:Artist",
@@ -20,13 +20,12 @@ public class CheckSkinningTests
             "Version:Test",
             "[Difficulty]",
             "StackLeniency:0.7",
+            "CircleSize:4",
             "[TimingPoints]",
             "0,500,4,2,0,100,1,0",
             "[HitObjects]",
             hitObject
         );
-
-    private const string DrumrollHitObject = "256,192,1000,2,0,B|300:200,1,50,0|0,0:0|0:0,0:0:0:0:";
 
     [Fact]
     public void CompleteHitburstSet_NoIssues()
@@ -35,15 +34,16 @@ public class CheckSkinningTests
             [("test.osu", BuildMinimalOsu())],
             [
                 "audio.mp3",
-                "taiko-hit0.png",
-                "taiko-hit100.png",
-                "taiko-hit100k.png",
-                "taiko-hit300.png",
-                "taiko-hit300k.png",
+                "mania-hit0.png",
+                "mania-hit50.png",
+                "mania-hit100.png",
+                "mania-hit200.png",
+                "mania-hit300.png",
+                "mania-hit300g.png",
             ]
         );
 
-        var issues = context.RunBeatmapSetCheck<CheckSkinning>();
+        var issues = context.RunBeatmapSetCheck<CheckSkinningMania>();
 
         Assert.Empty(issues);
     }
@@ -53,46 +53,26 @@ public class CheckSkinningTests
     {
         using var context = CheckTestContext.CreateFromOsuFiles(
             [("test.osu", BuildMinimalOsu())],
-            ["audio.mp3", "taiko-hit0.png"]
+            ["audio.mp3", "mania-hit0.png"]
         );
 
-        var issues = context.RunBeatmapSetCheck<CheckSkinning>();
+        var issues = context.RunBeatmapSetCheck<CheckSkinningMania>();
 
         Assert.Single(issues, issue => issue.level == Issue.Level.Problem);
         Assert.Contains("Hitburst", issues[0].message);
     }
 
     [Fact]
-    public void DrumrollElementsNotRequired_WithoutDrumrolls()
+    public void StageElements_AreNotBeatmapSkinnable_NoIssues()
     {
         using var context = CheckTestContext.CreateFromOsuFiles(
             [("test.osu", BuildMinimalOsu())],
-            [
-                "audio.mp3",
-                "taikobigcircle.png",
-                "taikobigcircleoverlay.png",
-                "taikohitcircle.png",
-                "taikohitcircleoverlay.png",
-            ]
+            ["audio.mp3", "mania-stage-left.png", "mania-stage-right.png"]
         );
 
-        var issues = context.RunBeatmapSetCheck<CheckSkinning>();
+        var issues = context.RunBeatmapSetCheck<CheckSkinningMania>();
 
         Assert.Empty(issues);
-    }
-
-    [Fact]
-    public void DrumrollElementsIncomplete_WithDrumrolls_EmitsProblem()
-    {
-        using var context = CheckTestContext.CreateFromOsuFiles(
-            [("test.osu", BuildMinimalOsu(DrumrollHitObject))],
-            ["audio.mp3", "taiko-roll-middle.png"]
-        );
-
-        var issues = context.RunBeatmapSetCheck<CheckSkinning>();
-
-        Assert.Single(issues, issue => issue.level == Issue.Level.Problem);
-        Assert.Contains("Hit object", issues[0].message);
     }
 
     [Fact]
@@ -102,15 +82,16 @@ public class CheckSkinningTests
             [("test.osu", BuildMinimalOsu())],
             [
                 "audio.mp3",
-                "taiko-hit0.png",
-                "taiko-hit100.png",
-                "taiko-hit100k.png",
-                "taiko-hit300.png",
-                "taiko-hit300k.jpg",
+                "mania-hit0.png",
+                "mania-hit50.png",
+                "mania-hit100.png",
+                "mania-hit200.png",
+                "mania-hit300.png",
+                "mania-hit300g.jpg",
             ]
         );
 
-        var issues = context.RunBeatmapSetCheck<CheckSkinning>();
+        var issues = context.RunBeatmapSetCheck<CheckSkinningMania>();
 
         Assert.Contains(issues, issue => issue.level == Issue.Level.Warning);
     }
